neural-networks/object-detection/conveyor-application-barcodes/requirements.txt
depthai == 3.0.0rc4
depthai-nodes
numpy
opencv-python
pyzbar
Pillow
zxing-cpp